server/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
sqlmodel==0.0.21
SQLAlchemy>=2.0.36
passlib==1.7.4
bcrypt==3.2.2
python-jose==3.3.0
python-multipart==0.0.9
google-generativeai==0.6.0
python-docx==1.1.2
python-pptx==0.6.23
jinja2==3.1.4

